services/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
pandas
python-dotenv
python-dateutil